+++ src/main/java/frc/robot/RobotContainer.java
@@ -20,7 +20,6 @@ import edu.wpi.first.wpilibj2.command.button.JoystickButton;
 import frc.robot.autochooser.chooser.AutoChooser;
 import frc.robot.autochooser.chooser.AutoChooser2024;
 import frc.robot.commands.RaiseArms;
-import frc.robot.commands.RatchetingMechanism;
 import frc.robot.commands.ReportErrorCommand;
 import frc.robot.commands.cannon.Shoot;
 import frc.robot.commands.cannon.StartFeeder;
@@ -134,7 +133,6 @@ public class RobotContainer {
         if (Constants.CLIMBER_DEBUG) {
             SmartShuffleboard.putCommand("Climber", "Climb", new StaticClimb(climber));
           SmartShuffleboard.putCommand("Climber", "RaiseArms", new RaiseArms(climber));
-          SmartShuffleboard.putCommand("Climber", "Lock Unlock Ratchet",new RatchetingMechanism(servo));
         }
         if (Constants.INTAKE_DEBUG){
             SmartShuffleboard.putCommand("Intake", "Start Intake", new StartIntake(intakeSubsystem,5));
